// Frontend/src/lib/types/typeTruckRegistration.tsx
// lib/submitVehicleForm.ts
export interface truckRegistration {  
    listingTitle: string;
          make: string;
          model: string;
          condition: string;
          inventoryLocation: string;
          transmission: string;
          fuelType: string;
          axleConfig: string;
          hotLocation: string;
          steering: string;
          year: number;
          mileage: number;
          price: number;
          description: string;
          category: string;
          color: string;
          engineSize: number;
          doors: number;
          seats: number;
          modelCode: string;
          auctionGrade: string;
          loadingCapacity: number;
          engineNumber: string;
          chassisNo: string;
          dimensionLength: number;
          dimensionWidth: number;
          dimensionHeight: number;
          carOptions: string[];
          vin: string;
          stockNumber: string;
          dealerStock: boolean;
          userId: string;
          imageBase64: string;
          imageUrl?: string;
  }
  
    